requirements.txt
flask
flask-cors
playwright
PyGithub
python-dotenv
